move debug implementation to tools dir
