+++ tools/debug/hack.go
